refactor(api-keys-ui): move create form to detail page, list-only config section
- ApiKeyProperties (new component) handles three modes:
  - key_id unset, no justCreated → create form (Name, Scope, Create/Cancel buttons)
  - key_id unset, justCreated set → "Save it now" view with the plaintext key,
    Copy-to-clipboard button, and "I've saved it" button. The latter sets
    key_id in the URL and switches into the detail view.
  - key_id set → existing detail/revoke view.
- ApiKeyDetailPage is now a thin wrapper rendering
  <div className="ApiKeyDetailPage"><ApiKeyProperties /></div>, matching the
  canonical <Foo>DetailPage / <Foo>Properties pattern (HoldingProperties /
  TransactionProperties / ConnectionProperties).
- ApiKeysSection (config page) drops the embedded create form. Now: header +
  list rows + a single "Add" button at the bottom that routes to
  api-key-detail with no key_id param (which triggers the create form mode).

Addresses hoiekim review (May 8 19:07) on PR #314.
