+++ src/client/components/ApiKeyProperties/index.css
@@ -1,0 +1,18 @@
+.ApiKeyProperties .apiKeyError {
+  color: var(--red);
+  font-size: 13px;
+}
+
+.ApiKeyProperties .apiKeyPrefix,
+.ApiKeyProperties .apiKeyPlaintext {
+  font-family: var(--font-mono, ui-monospace, SFMono-Regular, Menlo, monospace);
+  font-size: 13px;
+  word-break: break-all;
+  user-select: all;
+}
+
+.ApiKeyProperties .apiKeyCopyOnceTitle {
+  color: var(--red);
+  font-size: 13px;
+  font-weight: 600;
+}
