+++ src/client/pages/ApiKeyDetailPage/index.css
@@ -1,8 +1,3 @@
 div.ApiKeyDetailPage > div {
   padding: 0 10px;
 }
-
-.ApiKeyDetailPage .apiKeyPrefix {
-  font-family: var(--monoFont, monospace);
-  font-size: 13px;
-}
